# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Take home tech task
  version: v1.0.0
  description: We want you to create a REST API for Eagle Bank which conforms to this
    OpenAPI specification which allows a user to create, fetch, update and delete
    a bank account and deposit or withdraw money from the account. These will be stored
    as transactions against a bank account which be retrieved but not modified or
    deleted.
servers:
- url: http://localhost:8080
  description: Local development
...
paths:
...
  /v1/users:
    post:
      tags:
      - user
      description: Create a new user
      operationId: createUser
      requestBody:
        description: Create a new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '201':
          description: User has been created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid details supplied
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users/{userId}:
    get:
      tags:
      - user
      description: Fetch user by ID.
      operationId: fetchUserByID
      parameters:
      - name: userId
        in: path
        description: ID of the user
        required: true
        schema:
          type: string
          pattern: ^usr-[A-Za-z0-9]+$
      security:
      - bearerAuth: []
      responses:
        '200':
          description: The user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: The request didn't supply all the necessary data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'
        '401':
          description: Access token is missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The user is not allowed to access the transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - user
      description: Update user by ID.
      operationId: updateUserByID
      parameters:
      - name: userId
        in: path
        description: ID of the user
        required: true
        schema:
          type: string
          pattern: ^usr-[A-Za-z0-9]+$
      security:
      - bearerAuth: []
      requestBody:
        description: Update user details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        '200':
          description: The updated user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: The request didn't supply all the necessary data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'
        '401':
          description: Access token is missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The user is not allowed to access the transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - user
      description: Delete user by ID.
      operationId: deleteUserByID
      parameters:
      - name: userId
        in: path
        description: ID of the user
        required: true
        schema:
          type: string
          pattern: ^usr-[A-Za-z0-9]+$
      security:
      - bearerAuth: []
      responses:
        '204':
          description: The user has been deleted
        '400':
          description: The request didn't supply all the necessary data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'
        '404':
          description: User was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Access token is missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The user is not allowed to access the transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: A user cannot be deleted when they are associated with a bank
            account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CreateUserRequest:
      type: object
      example:
        password: password123
        address:
          town: town
          county: county
          postcode: postcode
          line3: line3
          line2: line2
          line1: line1
        phoneNumber: "+441234567890"
        name: user123
        email: user@example.com
      required:
        - name
        - address
        - phoneNumber
        - email
        - password
      properties:
        name:
          type: string
        address:
          type: object
          required:
            - line1
            - town
            - county
            - postcode
          properties:
            line1:
              type: string
            line2:
              type: string
            line3:
              type: string
            town:
              type: string
            county:
              type: string
            postcode:
              type: string
        phoneNumber:
          type: string
          pattern: ^\+[1-9]\d{1,14}$
        email:
          type: string
          format: email
        password:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: object
          required:
          - line1
          - town
          - county
          - postcode
          properties:
            line1:
              type: string
            line2:
              type: string
            line3:
              type: string
            town:
              type: string
            county:
              type: string
            postcode:
              type: string
        phoneNumber:
          type: string
          pattern: ^\+[1-9]\d{1,14}$
        email:
          type: string
          format: email
    UserResponse:
      type: object
      required:
      - id
      - name
      - address
      - phoneNumber
      - email
      - createdTimestamp
      - updatedTimestamp
      properties:
        id:
          type: string
          format: ^usr-[A-Za-z0-9]+$
        name:
          type: string
        address:
          type: object
          required:
          - line1
          - town
          - county
          - postcode
          properties:
            line1:
              type: string
            line2:
              type: string
            line3:
              type: string
            town:
              type: string
            county:
              type: string
            postcode:
              type: string
        phoneNumber:
          type: string
          pattern: ^\+[1-9]\d{1,14}$
        email:
          type: string
          format: email
        createdTimestamp:
          type: string
          format: date-time
        updatedTimestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
      - message
      properties:
        message:
          type: string
    BadRequestErrorResponse:
      type: object
      required:
      - message
      - details
      properties:
        message:
          type: string
        details:
          type: array
          items:
            type: object
            required:
            - field
            - message
            - type
            properties:
              field:
                type: string
              message:
                type: string
              type:
                type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT